Only create swapfile if required

+++ script/system/swapfile.sh
@@ -6,15 +6,18 @@ SWAP_FILE="/opt/muswap"
 SWAP_SIZE=$(GET_VAR "global" "settings/advanced/swapfile")
 
 CREATE_SWAP() {
-	LOG_INFO "$0" 0 "SWAPFILE" "Creating Swapfile"
-	dd if=/dev/zero of="$SWAP_FILE" bs=1M count="$SWAP_SIZE"
+	if [ "$2" -eq 1 ]; then
+		LOG_INFO "$1" 0 "SWAPFILE" "Creating Swapfile"
+		dd if=/dev/zero of="$SWAP_FILE" bs=1M count="$SWAP_SIZE"
+	fi
+	LOG_INFO "$1" 0 "SWAPFILE" "Mounting Swapfile"
 	chmod 0600 "$SWAP_FILE"
 	mkswap "$SWAP_FILE"
 	swapon "$SWAP_FILE"
 }
 
 PURGE_SWAP() {
-	LOG_INFO "$0" 0 "SWAPFILE" "Purging Swapfile"
+	LOG_INFO "$1" 0 "SWAPFILE" "Purging Swapfile"
 	swapoff "$SWAP_FILE"
 	rm -f "$SWAP_FILE"
 }
@@ -27,9 +30,11 @@ else
 		NEW_SIZE=$((SWAP_SIZE * 1024 * 1024))
 		if [ "$CURRENT_SIZE" -ne "$NEW_SIZE" ]; then
 			PURGE_SWAP "$0"
-			CREATE_SWAP "$0"
+			CREATE_SWAP "$0" "1"
+		else
+			CREATE_SWAP "$0" "0"
 		fi
 	else
-		CREATE_SWAP "$0"
+		CREATE_SWAP "$0" "1"
 	fi
 fi
